Refactoring module main.py - 15: I divided the module, which includes working with a cash register and a bank terminal.
